(PUP-10945) Rename HTTP APIs from master to server
Previously, we updated many 'master' terms to 'server'. However, we did
not update all of the "internal" references.

Puppet Server is also updating its terminology and loads the HTTP APIs
as part of its lifecycle. This updates the Puppet Server used HTTP APIs
to prefer 'server' to 'master'.
