+++ lib/puppet/network/http.rb
@@ -3,8 +3,11 @@ module Puppet::Network::HTTP
   HEADER_ENABLE_PROFILING = "X-Puppet-Profiling"
   HEADER_PUPPET_VERSION = "X-Puppet-Version"
 
-  MASTER_URL_PREFIX = "/puppet"
-  MASTER_URL_VERSIONS = "v3"
+  SERVER_URL_PREFIX = "/puppet"
+  SERVER_URL_VERSIONS = "v3"
+
+  MASTER_URL_PREFIX = SERVER_URL_PREFIX
+  MASTER_URL_VERSIONS = SERVER_URL_VERSIONS
 
   CA_URL_PREFIX = "/puppet-ca"
   CA_URL_VERSIONS = "v1"
