+++ lib/puppet/network/http/api/server.rb
@@ -1,0 +1,10 @@
+module Puppet
+  module Network
+    module HTTP
+      class API
+        module Server
+        end
+      end
+    end
+  end
+end
